feat: add authentication
On page load the application will now make a GET request to the /auth
endpoint of the API gateway. If an auth cookie is set, the user gets
signed in. If not, the sign in with google flow is being initiated.
Note: the expiration date for the auth cookie is set to 5 minutes after
the cookie is being set. This should probably change in the future for
the sake of a better user experience.