fix: don't interrupt smooth scrolling of to edit training
Because clicking on the edit CTA changes the editing props of the
component the scrolling of the element is interrupted. For now we use
the hacky way of using a timeout to delay the state update. In the
future the scrolling should probably be exchanged with an animation
using translateX.