feat: fuzzy suggestions
If no exercise starts with the current input we fallback to using
a fuzzy pattern matching to make a suggestion.

Note:
While it was fun playing around with the
`Levensthein Edit Distance` it doesn't produce good results in this use
case without modifications because it heavily considers the differnce in
string length.

Since the user is building up a string while typing the levensthein
distance delivers bad results.