feat: alternative input layer
In order to not trigger a layout shift because of the weird quirks
of having a bottom bar with an input element on mobile devices, we place
the input elemnt further to the top so the shift does not get triggered.
This also gives us space to add more functionality to the input in the
future.

For reference of the underlying problem:
https://github.com/bramus/viewport-resize-behavior/blob/main/explainer.md#the-icb