feat: remove headline from training format
The intention of the trainings format is to make it as easy as possible
to note down your workouts. Adding a headline is not neccessarily part
of that and can be down via an extra input field. This makes the
functionality more visible and eaasier to use.

I think in general it is a good idea to keep the trainings format as
simple as possible since we don't want a user to have to learn a weird
syntax in order to being able to use the application.